requirements.txt
asgiref==3.3.4
Django==3.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
gunicorn==20.1.0
Pillow==8.2.0
python-http-client==3.3.2
pytz==2021.1
sendgrid==6.7.0
sendgrid-python==0.1.1
sqlparse==0.4.1
starkbank-ecdsa==1.1.0

